#287 extract alert service to package
