+++ Packages/RuuviService/Sources/RuuviServiceAlert/AlertPersistence/AlertPersistence.swift
@@ -2,7 +2,6 @@ import Foundation
 import RuuviOntology
 
 protocol AlertPersistence {
-
     func alert(for uuid: String, of type: AlertType) -> AlertType?
     func register(type: AlertType, for uuid: String)
     func unregister(type: AlertType, for uuid: String)
