+++ station/Classes/Business/Background/Operations/WebTag/WebTagOperationsManager.swift
@@ -1,10 +1,11 @@
 import Foundation
 import RealmSwift
+import RuuviService
 
 class WebTagOperationsManager {
-
     var weatherProviderService: WeatherProviderService!
-    var alertService: AlertService!
+    var alertService: RuuviServiceAlert!
+    var alertHandler: AlertService!
     var webTagPersistence: WebTagPersistence!
 
     func alertsPullOperations() -> [Operation] {
@@ -14,19 +15,23 @@ class WebTagOperationsManager {
         for webTag in webTags {
             if alertService.hasRegistrations(for: webTag.uuid) {
                 if let location = webTag.location?.location {
-                    let operation = WebTagRefreshDataOperation(uuid: webTag.uuid,
-                                                               location: location,
-                                                               provider: webTag.provider,
-                                                               weatherProviderService: weatherProviderService,
-                                                               alertService: alertService,
-                                                               webTagPersistence: webTagPersistence)
+                    let operation = WebTagRefreshDataOperation(
+                        uuid: webTag.uuid,
+                        location: location,
+                        provider: webTag.provider,
+                        weatherProviderService: weatherProviderService,
+                        alertService: alertHandler,
+                        webTagPersistence: webTagPersistence
+                    )
                     operations.append(operation)
                 } else {
-                    let operation = CurrentWebTagRefreshDataOperation(uuid: webTag.uuid,
-                                                                      provider: webTag.provider,
-                                                                      weatherProviderService: weatherProviderService,
-                                                                      alertService: alertService,
-                                                                      webTagPersistence: webTagPersistence)
+                    let operation = CurrentWebTagRefreshDataOperation(
+                        uuid: webTag.uuid,
+                        provider: webTag.provider,
+                        weatherProviderService: weatherProviderService,
+                        alertService: alertHandler,
+                        webTagPersistence: webTagPersistence
+                    )
                     operations.append(operation)
                 }
             }
