+++ station/Classes/Business/Services/Alert/Impl/AlertServiceImpl+Process.swift
@@ -5,7 +5,6 @@ import RuuviOntology
 
 // MARK: - Process Physical Sensors
 extension AlertServiceImpl {
-
     // swiftlint:disable:next function_body_length
     func process(heartbeat ruuviTag: RuuviTagSensorRecord) {
         var isTriggered = false
@@ -52,14 +51,14 @@ extension AlertServiceImpl {
 
         let uuid = ruuviTag.ruuviTagId
         if let movementCounter = ruuviTag.movementCounter {
-            setMovement(counter: movementCounter, for: uuid)
+            ruuviAlertService.setMovement(counter: movementCounter, for: uuid)
         }
 
-        if hasRegistrations(for: uuid) {
+        if ruuviAlertService.hasRegistrations(for: uuid) {
             notify(uuid: uuid, isTriggered: isTriggered)
         }
 
-        if let macId = ruuviTag.macId?.mac, hasRegistrations(for: macId) {
+        if let macId = ruuviTag.macId?.mac, ruuviAlertService.hasRegistrations(for: macId) {
             notify(uuid: macId, isTriggered: isTriggered)
         }
     }
@@ -98,7 +97,7 @@ extension AlertServiceImpl {
             }
         }
 
-        if hasRegistrations(for: uuid) {
+        if ruuviAlertService.hasRegistrations(for: uuid) {
             notify(uuid: uuid, isTriggered: isTriggered)
         }
     }
@@ -137,7 +136,7 @@ extension AlertServiceImpl {
             }
         }
 
-        if hasRegistrations(for: identifier.value) {
+        if ruuviAlertService.hasRegistrations(for: identifier.value) {
             notify(uuid: identifier.value, isTriggered: isTriggered)
         }
     }
@@ -166,7 +165,7 @@ extension AlertServiceImpl {
     private func process(temperature: Temperature?,
                          alertType: AlertType,
                          identifier: Identifier) -> Bool {
-        if case .temperature(let lower, let upper) = alert(for: identifier.value, of: alertType),
+        if case .temperature(let lower, let upper) = ruuviAlertService.alert(for: identifier.value, of: alertType),
            let l = Temperature(lower),
            let u = Temperature(upper),
            let t = temperature {
@@ -191,7 +190,7 @@ extension AlertServiceImpl {
                          temperature: Temperature?,
                          alertType: AlertType,
                          identifier: Identifier) -> Bool {
-        if case .humidity(let lower, let upper) = alert(for: identifier.value, of: alertType),
+        if case .humidity(let lower, let upper) = ruuviAlertService.alert(for: identifier.value, of: alertType),
            let rh = humidity,
            let sh = Humidity(relative: rh.value, temperature: temperature) {
             let isLower = sh < lower
@@ -215,7 +214,7 @@ extension AlertServiceImpl {
                                  temperature: Temperature?,
                                  alertType: AlertType,
                                  identifier: Identifier) -> Bool {
-        if case .dewPoint(let lower, let upper) = alert(for: identifier.value, of: alertType),
+        if case .dewPoint(let lower, let upper) = ruuviAlertService.alert(for: identifier.value, of: alertType),
            let t = temperature,
            let rh = humidity,
            let sh = Humidity(relative: rh.value, temperature: t),
@@ -240,7 +239,7 @@ extension AlertServiceImpl {
     private func process(pressure: Pressure?,
                          alertType: AlertType,
                          identifier: Identifier) -> Bool {
-        if case .pressure(let lower, let upper) = alert(for: identifier.value, of: alertType),
+        if case .pressure(let lower, let upper) = ruuviAlertService.alert(for: identifier.value, of: alertType),
            let l = Pressure(lower),
            let u = Pressure(upper),
            let pressure = pressure {
@@ -262,7 +261,7 @@ extension AlertServiceImpl {
     }
 
     private func process(movement: AlertType, ruuviTag: RuuviTagSensorRecord) -> Bool {
-        if case .movement(let last) = alert(for: ruuviTag.ruuviTagId, of: movement),
+        if case .movement(let last) = ruuviAlertService.alert(for: ruuviTag.ruuviTagId, of: movement),
             let movementCounter = ruuviTag.movementCounter {
             let isGreater = movementCounter > last
             if isGreater {
