+++ station/Classes/Presentation/Modules/Dashboard/Charts/Assembly/Scroll/TagChartsScrollConfigurator.swift
@@ -26,7 +26,8 @@ class TagChartsScrollConfigurator {
         presenter.activityPresenter = r.resolve(ActivityPresenter.self)
         presenter.alertPresenter = r.resolve(AlertPresenter.self)
         presenter.mailComposerPresenter = r.resolve(MailComposerPresenter.self)
-        presenter.alertService = r.resolve(AlertService.self)
+        presenter.alertService = r.resolve(RuuviServiceAlert.self)
+        presenter.alertHandler = r.resolve(AlertService.self)
         presenter.foreground = r.resolve(BTForeground.self)
         presenter.background = r.resolve(BTBackground.self)
         presenter.feedbackEmail = r.property("Feedback Email")!
