+++ station/Classes/Presentation/Modules/TagSettings/Assembly/Table/TagSettingsTableConfigurator.swift
@@ -20,7 +20,7 @@ class TagSettingsTableConfigurator {
         presenter.foreground = r.resolve(BTForeground.self)
         presenter.background = r.resolve(BTBackground.self)
         presenter.calibrationService = r.resolve(CalibrationService.self)
-        presenter.alertService = r.resolve(AlertService.self)
+        presenter.alertService = r.resolve(RuuviServiceAlert.self)
         presenter.settings = r.resolve(RuuviLocalSettings.self)
         presenter.connectionPersistence = r.resolve(RuuviLocalConnections.self)
         presenter.pushNotificationsManager = r.resolve(PushNotificationsManager.self)
